spawn_yt_update.sh will be not longer overwriting existing update.sh files + scripts are more scripting friendly

+++ spawn_yt_update.sh
@@ -7,7 +7,13 @@ if [ -z $@ ]; then
     echo ""
     echo "Where link you can pass every link able to pass to youtube-dl. It mean not only YT links!"
     echo "It will spawn your "yt repo file", just start ./update and when you have yt in your path, it will start downloading repo. Later it will update it."
-    exit 0
+    exit 1
+fi
+
+if [ -f update.sh ]; then
+    echo "update.sh allready exists, you can not spawn new update.sh file"
+    echo "Appending new links to existing spawner will be added later"
+    exit 1
 fi
 
 echo "#!/bin/bash" > update.sh
@@ -19,3 +25,4 @@ echo "while true ; do yt \"$1\" && break ; done" >> update.sh
 chmod +x update.sh
 stat update.sh
 
+exit 0
